Do not overwrite Dockerfile.test
When building test image, use a modified dockerfile in temporary
directory instead of changing the checked in Dockerfile.test.

Bumped Go to 1.8

Moved build.go from its own package to be a part of the wrench package
because I could not for the life of me get Go to use the changed code
when it was its own package. I can very well see the wrench commands not
being their own packages, so this is a start.

Fixes #26